refactor: Moved label edit controls to a separate component
